training/requirements.txt
torch
torchvision
torchaudio
mlflow==2.17.2
PyYAML
numpy
onnx
onnxruntime
pandas
scikit-learn
tqdm
